# backend/public/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}}},"responses":{"201":{"description":""},"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Token"}}}}}},"tags":["Auth"]}},"/api/auth/register":{"post":{"operationId":"AuthController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupDTO"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}}},"tags":["Auth"]}},"/api/auth/refresh-token":{"post":{"operationId":"AuthController_refreshToken","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenInput"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Token"}}}}}},"tags":["Auth"]}},"/api/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/api/ocr/ekyc":{"post":{"operationId":"OcrController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOcrDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/ocr/vehicle-registrations":{"post":{"operationId":"OcrController_vehicleRegistrations","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOcrDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/api/ocr":{"get":{"operationId":"OcrController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/api/ocr/{id}":{"get":{"operationId":"OcrController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}},"patch":{"operationId":"OcrController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOcrDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}},"delete":{"operationId":"OcrController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}}},"info":{"title":"Nestjs FTW","description":"The nestjs API description","version":"1.5","contact":{}},"tags":[],"servers":[{"url":"http://localhost:8001"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDTO":{"type":"object","properties":{"email":{"type":"string","default":"a@gmail.com"},"password":{"type":"string","default":"12345678"}},"required":["email","password"]},"Token":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}},"required":["accessToken","refreshToken"]},"SignupDTO":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"firstname":{"type":"string"},"lastname":{"type":"string"}},"required":["email","password"]},"RefreshTokenInput":{"type":"object","properties":{"token":{"type":"string","default":"12345678"}},"required":["token"]},"CreateOcrDto":{"type":"object","properties":{"image":{"type":"string"}},"required":["image"]},"UpdateOcrDto":{"type":"object","properties":{"image":{"type":"string"}}}}}}